node: Move the analyze code to the stats module
