+++ node/src/manager/commands/mod.rs
@@ -1,4 +1,3 @@
-pub mod analyze;
 pub mod assign;
 pub mod chain;
 pub mod config;
